R# Move namespace and change name of class
